BCSS-20562 - BCSS - Playwright - Review pytest markers (#77)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description
This '@pytest.mark.smoke' added for all missing functions in 'test_'
files under 'tests' folder as these are critical functions which will be
considered during the smoke Test run.
This '@pytest.mark.smokescreen' added for all missing functions in 'test_compartment_'
files under 'tests --> Smokescreen' folder as these are critical functions which will be
considered during the smokescreen Test run of compartments 1 to 6.
This '@pytest.mark.vpn_required' is added only for missing functions
which uses Database.

<!-- Describe your changes in detail. -->
## Context
Just it is a code review for pytest markers & No code change is
implemeted.

<!-- Why is this change required? What problem does it solve? -->
## Type of changes
This line of syntax is added as it is required in order to identify &
invoke all the critical functionalities when we are doing Smoke Test run
of Compartments (1 to 6).

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->
- [X] Refactoring (non-breaking change)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist
<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->
- [X] I am familiar with the [contributing
guidelines](https://github.com/nhs-england-tools/playwright-python-blueprint/blob/main/CONTRIBUTING.md)
- [X] I have followed the code style of the project
- [ ] I have added tests to cover my changes (where appropriate)
- [ ] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

## Sensitive Information Declaration
To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [X] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.