BSS2-2306 (#15)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

Wrap interpolated markers value in double quotes so that it can accept a
string with spaces as an argument.

Rename 'tag' to 'markers' to reflect that the values are referred to as
markers in Pytest DSL.

## Context

The 'release' and 'smoke' markers are passed in to the Playwright image
from the Jenkins pipeline depending on whether the bs-select-app branch
is named 'release*'

Note: pytest markers are entered in the format <marker> and/or/not
<marker>. Because of the spaces the value has to be passed in double
quotes, like so "smoke or release". Additional double quotes are
required in bs-select-playwright when the passed value is populated.

See
[this](https://gitlab.mgmt.texasplatform.uk/bs-select/bs-select-app/-/merge_requests/934)
`bs-select-app` merge request for more context.

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [x] Refactoring (non-breaking change)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [x] I am familiar with the [contributing
guidelines](https://github.com/nhs-england-tools/playwright-python-blueprint/blob/main/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [ ] I have added tests to cover my changes (where appropriate)
- [ ] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [ x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.